+++ lib/src/duration_extension.dart
@@ -1,8 +1,11 @@
+import 'package:flutter/material.dart';
+import 'gen_l10n/simple_localizations.dart';
+
 import 'duration_formatters/duration_formatters.dart';
 
 extension DurationExtensions on Duration {
   /// Format duration to readable string
-  /// you can choose one of our formatters `DefaultDurationFormatter` or `ArabicEnglishDurationFormatter`
+  /// you can choose one of our formatters `DefaultDurationFormatter` or `TranslatedDurationFormatter`
   /// or you can implement `DurationFormatter` and create your own
   String format({
     int maxParts = 2,
@@ -15,6 +18,8 @@ extension DurationExtensions on Duration {
     delta -= years * (2592000 * 12);
     var months = (delta / 2592000).floor();
     delta -= months * 2592000;
+    var weeks = (delta / 86400).floor() ~/ 7;
+    delta -= weeks * 7 * 86400;
     var days = (delta / 86400).floor();
     delta -= days * 86400;
     var hours = (delta / 3600).floor() % 24;
@@ -37,6 +42,14 @@ extension DurationExtensions on Duration {
         text = formatter.getMonths(months);
       else
         text += formatter.getSeparator() + formatter.getMonths(months);
+      parts++;
+      if (parts >= maxParts) return text;
+    }
+    if (weeks > 0) {
+      if (text == null)
+        text = formatter.getWeeks(weeks);
+      else
+        text += formatter.getSeparator() + formatter.getWeeks(days);
       parts++;
       if (parts >= maxParts) return text;
     }
@@ -73,5 +86,92 @@ extension DurationExtensions on Duration {
       if (parts >= maxParts) return text;
     }
     return text ?? "";
+  }
+
+  ///Localize duration useing TranslatedDurationFormatter formatter
+  ///You have to add `SimpleLocalizations.delegate` in your `localizationsDelegates`
+  /// ```
+  /// import 'gen_l10n/simple_localizations.dart';
+  ///
+  /// return MaterialApp(
+  ///   localizationsDelegates: SimpleLocalizations.localizationsDelegates,
+  ///   supportedLocales: SimpleLocalizations.supportedLocales,
+  ///   home: MyApplicationHome(),
+  /// );
+  /// ```
+  String localeFormat(
+    BuildContext context, {
+    int maxParts = 2,
+    int secondsToShowNow = 0,
+  }) {
+    var simpleLocalization = SimpleLocalizations.of(context);
+    if (simpleLocalization == null)
+      throw FlutterError(
+          "You have to add `SimpleLocalizations.delegate` in your `localizationsDelegates`");
+    return format(
+      maxParts: maxParts,
+      secondsToShowNow: 0,
+      formatter: TranslatedDurationFormatter(context),
+    );
+  }
+
+  /// count down this duration with 1 microsecond
+  Duration countDownMicroseconds() {
+    return this - Duration(microseconds: 1);
+  }
+
+  /// count down this duration with 1 millisecond
+  Duration countDownMilliseconds() {
+    return this - Duration(milliseconds: 1);
+  }
+
+  /// count down this duration with 1 seconds
+  Duration countDownSeconds() {
+    return this - Duration(seconds: 1);
+  }
+
+  /// count down this duration with 1 minute
+  Duration countDownMinutes() {
+    return this - Duration(minutes: 1);
+  }
+
+  /// count down this duration with 1 hour
+  Duration countDownHours() {
+    return this - Duration(minutes: 1);
+  }
+
+  /// count down this duration with 1 day
+  Duration countDownDays() {
+    return this - Duration(days: 1);
+  }
+
+  /// count up this duration with 1 microsecond
+  Duration countUpMicroseconds() {
+    return this + Duration(microseconds: 1);
+  }
+
+  /// count up this duration with 1 millisecond
+  Duration countUpMilliseconds() {
+    return this + Duration(milliseconds: 1);
+  }
+
+  /// count up this duration with 1 seconds
+  Duration countUpSeconds() {
+    return this + Duration(seconds: 1);
+  }
+
+  /// count up this duration with 1 minute
+  Duration countUpMinutes() {
+    return this + Duration(minutes: 1);
+  }
+
+  /// count up this duration with 1 hour
+  Duration countUpHours() {
+    return this + Duration(minutes: 1);
+  }
+
+  /// count up this duration with 1 day
+  Duration countUpnDays() {
+    return this + Duration(days: 1);
   }
 }