+++ lib/src/duration_formatters/duration_formatters.dart
@@ -2,9 +2,10 @@ library duration_formatters;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import '../boolean_extensions.dart';
+import '../gen_l10n/simple_localizations.dart';
+import 'package:simple/src/gen_l10n/simple_localizations.dart';
 part 'default_formatter.dart';
-part 'arabic_english_formatter.dart';
+part 'translated_formatter.dart';
 
 abstract class DurationFormatter {
   String getSeparator();
@@ -12,6 +13,7 @@ abstract class DurationFormatter {
   String getMonths(int months);
 
   String getDays(int days);
+  String getWeeks(int weeks);
 
   String getHours(int hours);