+++ lib/src/timer_extension.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 extension TimerExtensions on Timer {
+  /// return `T` when timer is still active and not null else return null
   T? onActive<T>(T callback) {
     if (this.isActive)
       return callback;
@@ -8,6 +9,7 @@ extension TimerExtensions on Timer {
       return null;
   }
 
+  /// return `T` when timer is still active or is null else return null
   T? onActiveOrNull<T>(T callback) {
     if (this.isActive)
       return callback;
@@ -15,6 +17,7 @@ extension TimerExtensions on Timer {
       return null;
   }
 
+  /// return `T` when timer is not active or is null else return null
   T? onNotActiveOrNull<T>(T callback) {
     if (!this.isActive)
       return callback;
@@ -22,6 +25,7 @@ extension TimerExtensions on Timer {
       return null;
   }
 
+  /// return `T` when timer is not active and not null else return null
   T? onNotActive<T>(T callback) {
     if (!this.isActive)
       return callback;